rename func to bpf
