+++ runtime/wasm-bpf-rs/src/bpf/map_operate.rs
@@ -12,7 +12,7 @@ use libbpf_rs::libbpf_sys::{
 };
 use log::{debug, error};
 
-use crate::{ensure_enough_memory, func::EINVAL, state::CallerType, utils::CallerUtils};
+use crate::{ensure_enough_memory, bpf::EINVAL, state::CallerType, utils::CallerUtils};
 
 use super::WasmPointer;
